+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/IndexOptions.java
@@ -169,7 +169,7 @@ public class IndexOptions {
     public static final String CLEAR_WHEN_ZERO = "clearWhenZero";
 
     /**
-     * Size of the portion of the grouping keys enumerated after the extrema by {@link IndexTypes#PERMUTED_MIN} and {@link IndexTypes#PERMUTED_MAX} indexes.
+     * Size of the portion of the grouping keys enumerated after the extrema by {@link IndexTypes#MIN} and {@link IndexTypes#MAX} indexes.
      */
     public static final String PERMUTED_SIZE_OPTION = "permutedSize";